Fixup tests

+++ Tests/TestQHelpers.swift
@@ -6,28 +6,28 @@ final class TestQHelpers: XCTestCase {
         let namegate: SequentialObjectBlockingNameGate = .init()
 
         // 0,0 with no history should go through.
-        XCTAssert(namegate.handle(groupId: UInt32.random(in: 0...UInt32.max),
+        XCTAssert(namegate.handle(groupId: UInt64.random(in: 0...UInt64.max),
                                   objectId: 0,
                                   lastGroup: nil,
                                   lastObject: nil))
         // Non 0 object ID with no history should not go through.
-        XCTAssertFalse(namegate.handle(groupId: UInt32.random(in: 0...UInt32.max),
-                                       objectId: UInt16.random(in: 1...UInt16.max),
+        XCTAssertFalse(namegate.handle(groupId: UInt64.random(in: 0...UInt64.max),
+                                       objectId: UInt64.random(in: 1...UInt64.max),
                                        lastGroup: nil,
                                        lastObject: nil))
 
         // Sequential object in a group should go through.
         do {
-            let currentGroup: UInt32 = .random(in: 0...UInt32.max)
-            let currentObject: UInt16 = .random(in: 0...UInt16.max - 1)
+            let currentGroup: UInt64 = .random(in: 0...UInt64.max)
+            let currentObject: UInt64 = .random(in: 0...UInt64.max - 1)
             XCTAssert(namegate.handle(groupId: currentGroup,
                                       objectId: currentObject + 1,
                                       lastGroup: currentGroup,
                                       lastObject: currentObject))
             // Sequential objects in a different group should not go through.
-            var differentGroup: UInt32
+            var differentGroup: UInt64
             repeat {
-                differentGroup = .random(in: 0...UInt32.max)
+                differentGroup = .random(in: 0...UInt64.max)
             }
             while differentGroup == currentGroup
             XCTAssertFalse(namegate.handle(groupId: differentGroup,
@@ -38,9 +38,9 @@ final class TestQHelpers: XCTestCase {
 
         // Non sequential object ID in a group should not go through.
         do {
-            let currentGroup: UInt32 = .random(in: 0...UInt32.max)
-            let currentObject: UInt16 = .random(in: 0...UInt16.max - 2)
-            let new: UInt16 = .random(in: currentObject+2...UInt16.max)
+            let currentGroup: UInt64 = .random(in: 0...UInt64.max)
+            let currentObject: UInt64 = .random(in: 0...UInt64.max - 2)
+            let new: UInt64 = .random(in: currentObject+2...UInt64.max)
             XCTAssertFalse(namegate.handle(groupId: currentGroup,
                                            objectId: new,
                                            lastGroup: currentGroup,
@@ -49,17 +49,17 @@ final class TestQHelpers: XCTestCase {
 
         // Incrementing group jumps should go through when their objectId is 0.
         do {
-            let currentGroup: UInt32 = .random(in: 0...UInt32.max - 2)
-            let nextGroup: UInt32 = .random(in: currentGroup+1...UInt32.max-1)
+            let currentGroup: UInt64 = .random(in: 0...UInt64.max - 2)
+            let nextGroup: UInt64 = .random(in: currentGroup+1...UInt64.max-1)
             XCTAssert(namegate.handle(groupId: nextGroup, objectId: 0, lastGroup: currentGroup, lastObject: 0))
             // Even when it jumps non-sequentially.
             XCTAssert(namegate.handle(groupId: nextGroup + 1, objectId: 0, lastGroup: currentGroup, lastObject: 0))
             // Group jumps should not go through when their objectId is not 0.
-            let nonZeroObject: UInt16 = .random(in: 1...UInt16.max)
+            let nonZeroObject: UInt64 = .random(in: 1...UInt64.max)
             XCTAssertFalse(namegate.handle(groupId: nextGroup,
                                            objectId: nonZeroObject,
                                            lastGroup: currentGroup,
-                                           lastObject: UInt16.random(in: 0...UInt16.max)))
+                                           lastObject: UInt64.random(in: 0...UInt64.max)))
         }
 
         // Example series.
